requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
holidays>=0.45
streamlit-aggrid>=0.3.4
